requirements.txt
opencv-python>=4.9
mediapipe>=0.10.9
sounddevice>=0.4.6
numpy>=1.24
faster-whisper>=1.0.0
openai>=1.12
python-dotenv>=1.0
